Description:fix cppcheck fail
Match-id-416a4cdd5876d1a122856a0d445e37a82796bc11

+++ interfaces/inner_api/dlp_fuse/test/unittest/src/dlp_fuse_test.cpp
@@ -841,7 +841,7 @@ HWTEST_F(DlpFuseTest, AddDlpLinkFile008, TestSize.Level1)
     // get txtSize
     ASSERT_EQ(read(g_dlpFileFd, readBuf, 16), 16);
     uint32_t* hdr = reinterpret_cast<uint32_t*>(readBuf);
-    ASSERT_EQ(hdr[3], 4);
+    ASSERT_EQ(static_cast<int>(hdr[3]), 4);
 }
 
 /**
